Handle file and image fields by referencing files by name.
